requirements.txt
djangorestframework
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
idna==2.8
lxml==4.4.2
pdfminer.six==20181108
pdfplumber==0.5.14
Pillow==6.2.1
pycryptodome==3.9.4
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sortedcontainers==2.1.0
sqlparse==0.3.0
unicodecsv==0.14.1
urllib3==1.25.7
Wand==0.5.8
